demo_k8_jobs_bqs_gcs/requirements.txt
google-cloud-bigquery-storage
google-cloud-bigquery
google-cloud-storage
fastavro